add: Repeater can now repeat sets of bits.
Repeater chip can now accept more than one input, and when doing so, will
repeat sets of bits. The old behavior of repeating the first input on
all outputs is still preserved, as long as the chip only has one input.
Note that as long as all existing repeater chips have only 1 input,
this is backwards compatible with them. Behavior will not change.

Example:
Inputs:  1234
Outputs: 12345678

Prior behavior:
Input 1 is repeated on all outputs. Inputs 2-4 are ignored.

New behavior:
Input 1 is repeated on outputs 1 and 5. Input 2 is repeated on outputs
2 and 6, and so on.