add: dregister chips can now be backed by ram
The dregister chip can now be backed by any ram object, by providing
it's id as the second sign argument, in the format "$<ram_id>".
The third sign argument specifies the address to look in (defaults to 0).
The output of the register will be automatically updated when the
relevant ram address updates, and triggering the clock pin will
update the relevant ram address. Other than the dregister being
backed by ram, it functions exactly the same. If no ram id is
specified, the dregister functions as normal.