add: added ram watcher chip
Added the ram watcher chip. The watcher chip takes 1 input and 1 output.
It also takes 1 or 2 arguments. The first is mandatory and specifies the
ram to watch, while the 2 specifies the ram address to watch.
The input functions as a disable/enable pin. If it is off, the chip
does not function. The output pin will send a 1-tick pulse when the
specified ram address in the specified ram id is changed. (If no address
is given, the entire ram is watched.) This is useful for triggering the
clock pin of a ram-backed display when it's ram-page is determined by
a ram-backed dregister.