add repeated trigger mode setting to pulse circuit
The pulse circuit had (most likely) undesired behavior when the Circuit
was triggered while a pulse was active. The subsequent triggers would
unconditionally schedule tasks to turn off the signal after the pulse
length. So if another pulse started after the first pulse ended, but there
were additional triggers during the initial pulse, then the second pulse
would turn off early.

This commit adds a "repeated trigger" parameter to the pulse circuit.
The 3rd parameter is optional and defaults to "legacy" for the original
behavior. A value of "ignore" causes the circuit to ignore any triggers
while the pulse is active. A value of "extend" causes the circuit to cancel
the original task that would turn off the pulse and schedule a new one
(for the original pulse length).